refactor: 라우트 파일이랑 스타일 파일을 한 폴더에 모아놓음
